Move argparse and recorder to headers
